agentic/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-starter"
version = "0.2.0"
description = "Schema-first, tool-centric, Temporal-orchestrated coding agent starter"
authors = [{ name = "Rob" }]
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.8",
  "pydantic-settings>=2.3",
  "pydantic-ai-slim[openai]>=0.0.16",
  "temporalio>=1.6",
  "httpx>=0.27",
  "typer>=0.12",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "langfuse>=3.2.5",
  "pytest>=8.3",
  "ruff>=0.5",
  "mypy>=1.10",
  "bandit>=1.7",
  "unidiff>=0.7.5",
  "tree-sitter-languages>=1.10",
  "tree-sitter>=0.23",
]

[project.scripts]
agentic = "agentic.orchestration.client_cli:app"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"

[tool.setuptools]
packages = ["agentic"]
